Fix by Stefan Van Dyck
When scaling our figures the fact we cant pass floats as coords resulted
in a lot of artefacts.

I've found a very easy fix to use floats as coords when adding objects
to a powerpoint slide.

on line 530 in genpptx.js:
add a Math.round() function to the realNum var in parseSmartNumber.

That way we can place object more accuratly.