Remove temporary resources just before closing
There seems to be a race condition where the sample[n].xlsx files are 
removed before the archiver has a chance to append/include them.

This causes the error 'no such file or directory sample[n].xlsx'.

This PR moves the removal of temporary resources to much later in the 
build process - after all the resources have been added.