feat: pPTX workaround for using placeholders + custom sizes + custom fills: disableFillSettings
In case that you are adding a place holder field to a slide and the original field in the layout has
custom fill settings, and also you want to change the size and position of the field then you must
use the disableFillSettings = true setting. This is a workaround for this case.